% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-mpg.R
\name{plot_mpg}
\alias{plot_mpg}
\title{Plot the mpg dataset's displ and hwy variable}
\usage{
plot_mpg(colour_var = NULL, facet_var = NULL)
}
\arguments{
\item{colour_var}{A variable name as a string to use as a
mapping in \code{\link[ggplot2:aes]{ggplot2::aes()}} for the colour aesthetic, or \code{NULL} to suppress.}

\item{facet_var}{A variable name as a string to use as in
\code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}} for the colour aesthetic, or \code{NULL} to skip facetting.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object.
}
\description{
This funciton provides a way for plotting mpg dataset's displ and hwy
with different facet and colour variable using \link{ggplot2} package.
}
\examples{
plot_mpg("class", "drv")
plot_mpg(NULL, "drv")
plot_mpg(NULL, NULL)
}
