% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-mpg-string.R
\name{plot_mpg_string}
\alias{plot_mpg_string}
\title{Plot the mpg dataset using string variable specification}
\usage{
plot_mpg_string(colour_var = NULL, facet_var = NULL)
}
\arguments{
\item{colour_var}{A variable name as a string to use as a
\link[ggplot2:aes]{mapping} for the colour aesthetic,
or \code{NULL} to suppress.}

\item{facet_var}{A variable name as a string to use in
\code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}}, or \code{NULL} to skip facetting.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object.
}
\description{
Plot the mpg dataset using string variable specification
}
\examples{
plot_mpg_string("class", "manufacturer")

}
