% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-mpg.R
\name{plot_mpg}
\alias{plot_mpg}
\title{Plot the mpg dataset}
\usage{
plot_mpg(colour_var = NULL, facet_var = NULL)
}
\arguments{
\item{colour_var}{A variable or expression to use as a
\link[ggplot2:aes]{mapping} for the colour aesthetic,
or \code{NULL} to suppress.}

\item{facet_var}{A variable or expression to use in
\code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}}, or \code{NULL} to skip facetting.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object.
}
\description{
Plot the mpg dataset
}
\examples{
plot_mpg(class, manufacturer)

}
